Made OSG2 compile on Mac OS X 10.5 Leopard
git-svn-id: https://opensg.vrsource.org/svn/trunk@1019 4683daeb-ad0f-0410-a623-93161e962ae5

+++ Source/WindowSystem/CoreGL/testWindowCoreGL.cpp
@@ -375,7 +375,7 @@ int main (int argc, char **argv)
         (CGLPixelFormatAttribute)0
     };
     CGLPixelFormatObj pixelFormatObj;
-    long numPixelFormats;
+    GLint numPixelFormats;
     CGLChoosePixelFormat(attribs, &pixelFormatObj, &numPixelFormats); 
 
     CGLContextObj contextObj;
